feat: 승패 여부를 반환하는 로직 개선
