+++ src/main/java/blackjack/dto/DealerResultsDto.java
@@ -1,12 +1,12 @@
 package blackjack.dto;
 
-import static java.util.stream.Collectors.*;
 import static java.util.stream.Collectors.counting;
 import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
 
 import blackjack.domain.participant.Dealer;
-import blackjack.domain.GameResult;
 import blackjack.domain.participant.Players;
+import blackjack.domain.result.GameResult;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
@@ -22,7 +22,7 @@ public class DealerResultsDto {
     public static DealerResultsDto of(Players players, Dealer dealer) {
         Map<GameResult, Long> results = players.getValue()
                 .stream()
-                .collect(groupingBy(player -> dealer.decideResult(player.getTotalScore()),
+                .collect(groupingBy(player -> dealer.decideResult(player),
                         () -> new EnumMap<>(GameResult.class), counting()));
 
         List<DealerResultDto> value = results.keySet()
