+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,16 +1,24 @@
 package blackjack.domain.participant;
 
-import static blackjack.domain.CardsTestDataGenerator.*;
-import static blackjack.domain.card.Denomination.*;
-import static blackjack.domain.GameResult.*;
-import static blackjack.domain.card.Suit.*;
+import static blackjack.domain.CardsTestDataGenerator.generateBlackjack;
+import static blackjack.domain.CardsTestDataGenerator.generateCards;
+import static blackjack.domain.CardsTestDataGenerator.generateTotalScoreGraterThan17Cards;
+import static blackjack.domain.CardsTestDataGenerator.generateTotalScoreNotMoreThan16Cards;
+import static blackjack.domain.card.Denomination.ACE;
+import static blackjack.domain.card.Denomination.FIVE;
+import static blackjack.domain.card.Denomination.JACK;
+import static blackjack.domain.card.Denomination.KING;
+import static blackjack.domain.card.Suit.HEART;
+import static blackjack.domain.card.Suit.SPADE;
+import static blackjack.domain.result.GameResult.LOSE;
+import static blackjack.domain.result.GameResult.WIN;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import blackjack.domain.CardsArgumentsProvider;
-import blackjack.domain.GameResult;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
+import blackjack.domain.result.GameResult;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -75,7 +83,7 @@ public class PlayerTest {
 
         Player player = new Player("sudal", generateBlackjack());
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(WIN);
     }
@@ -91,7 +99,7 @@ public class PlayerTest {
         player.append(new Card(KING, HEART));
         player.append(new Card(JACK, SPADE));
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(LOSE);
     }
@@ -104,7 +112,7 @@ public class PlayerTest {
         player.append(new Card(KING, SPADE));
         player.append(new Card(JACK, SPADE));
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(GameResult.LOSE);
     }
@@ -115,7 +123,7 @@ public class PlayerTest {
         Dealer dealer = new Dealer(generateTotalScoreNotMoreThan16Cards());
         Player player = new Player("sudal", generateTotalScoreGraterThan17Cards());
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(GameResult.WIN);
     }
@@ -126,7 +134,7 @@ public class PlayerTest {
         Dealer dealer = new Dealer(generateTotalScoreGraterThan17Cards());
         Player player = new Player("sudal", generateTotalScoreNotMoreThan16Cards());
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(GameResult.LOSE);
     }
@@ -137,7 +145,7 @@ public class PlayerTest {
         Dealer dealer = new Dealer(generateCards());
         Player player = new Player("sudal", generateCards());
 
-        GameResult gameResult = player.decideResult(dealer.getTotalScore());
+        GameResult gameResult = player.decideResult(dealer);
 
         assertThat(gameResult).isEqualTo(GameResult.TIE);
     }
